requirements.txt
# requirements.txt - Complete Dependencies for DeepBTC Project

# Core Data Manipulation
pandas==2.1.4
numpy==1.26.2

# Market Data Acquisition
ccxt==4.1.85                    # Cryptocurrency exchange data
websocket-client==1.7.0         # Real-time WebSocket streams

# Technical Analysis
pandas-ta==0.3.14b0             # Technical indicators library
ta-lib==0.4.28                  # Alternative TA library (optional, requires binary)

# Alternative Market Data (No API key required)
yfinance==0.2.33                # Yahoo Finance data (stocks, indices, commodities)

# HTTP Requests
requests==2.31.0                # API calls to external sources

# Data Visualization (for exploratory analysis)
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Machine Learning / Deep Learning
tensorflow==2.15.0              # or tensorflow-gpu for GPU support
keras==2.15.0
scikit-learn==1.3.2
xgboost==2.0.3                  # Gradient boosting (ensemble models)

# Model Optimization
optuna==3.5.0                   # Hyperparameter tuning
keras-tuner==1.4.6              # Keras-specific tuning

# Feature Engineering Helpers
scikit-optimize==0.9.0          # Bayesian optimization
featuretools==1.29.0            # Automated feature engineering (optional)

# Time Series Specific
statsmodels==0.14.1             # Statistical models
prophet==1.1.5                  # Facebook's time series library (optional)

# Model Persistence
joblib==1.3.2                   # Model serialization
h5py==3.10.0                    # HDF5 for Keras models

# Utilities
python-dotenv==1.0.0            # Environment variable management
tqdm==4.66.1                    # Progress bars
colorama==0.4.6                 # Colored terminal output

# Jupyter Notebooks (for analysis)
jupyter==1.0.0
ipykernel==6.27.1
notebook==7.0.6

# Testing (optional but recommended)
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality (optional)
black==23.12.1                  # Code formatter
flake8==7.0.0                   # Linter
mypy==1.8.0                     # Type checker

# Documentation (optional)
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Production Deployment (future)
fastapi==0.109.0                # API framework for model serving
uvicorn==0.27.0                 # ASGI server
redis==5.0.1                    # Caching/message broker
celery==5.3.4                   # Task queue for async jobs

# Monitoring (future)
prometheus-client==0.19.0       # Metrics
sentry-sdk==1.39.2              # Error tracking

# Notes:
# 1. For GPU support, install tensorflow-gpu instead of tensorflow
# 2. TA-Lib requires binary installation: https://github.com/mrjbq7/ta-lib
# 3. Some libraries (prophet, ta-lib) may need additional system dependencies
# 4. For production, pin all versions exactly as shown